added assault hospitalizations indicator
