+++ internal_review/quality_of_life/borough/assault_hospitalizations.csv
@@ -1,0 +1,6 @@
+borough,safety_assault_hospital_per100k
+BK,42.92
+BX,80.9
+MN,34.04
+QN,29.76
+SI,42.1
