+++ internal_review/quality_of_life/citywide/assault_hospitalizations.csv
@@ -1,0 +1,2 @@
+citywide,safety_assault_hospital_per100k
+citywide,43.97
