+++ internal_review/quality_of_life/puma/assault_hospitalizations.csv
@@ -1,0 +1,56 @@
+puma,safety_assault_hospital_per100k
+03701,20.64
+03702,70.24
+03703,30.73
+03704,57.4
+03705,126.36
+03706,73.95
+03707,76.44
+03708,90.82
+03709,59.2
+03710,170.44
+03801,35.26
+03802,49.0
+03803,83.82
+03804,88.79
+03805,10.0
+03806,12.91
+03807,30.34
+03808,30.12
+03809,36.13
+03810,9.27
+03901,14.98
+03902,34.85
+03903,72.26
+04001,26.58
+04002,40.84
+04003,67.33
+04004,39.15
+04005,22.92
+04006,72.68
+04007,105.24
+04008,82.56
+04009,37.72
+04010,74.07
+04011,62.99
+04012,20.6
+04013,16.87
+04014,21.95
+04015,32.99
+04016,15.03
+04017,19.23
+04018,51.78
+04101,65.41
+04102,25.06
+04103,15.36
+04104,
+04105,20.68
+04106,19.85
+04107,31.29
+04108,
+04109,14.13
+04110,18.32
+04111,34.19
+04112,59.76
+04113,37.58
+04114,53.05
