+++ resources/quality_of_life/pedestrian_injuries.csv
